Fixes for Python 3 and a minor display improvement
Made the 'default' namespace not print on the context line since it takes up a lot of valuable room.
Added support for more modern versions of kubernetes_py.
Fixed a warning issue caused by kubernetes_py using yaml.load instead of yaml.safe_load which ends up printing on every load without suppressing the error.  I'm submitting a PR for that to kubernetes_py as well to fix it.